Added best score
Hi guys. For some reasons my git was not working so I couldn't see all
the changes since 21/06. Sorry about that. I see John added the score so
I added the best score which is a feature that records the best score. I
also added bestScoreLable so it can display the best score to the player
when he/she is playing the game. I don't know this is working or not
because I can't build the latest version of the game in Visual Studio.